fixed compilation with gcc4.3 broken pfnet
git-svn-id: https://piggledy.org/svn/peerfuse/trunk@952 f7b84f85-2e04-4de5-9e99-c4612dc0539c

+++ common/file_chunk_desc.h
@@ -67,7 +67,7 @@ public:
 		off_t end_off = MIN(GetEndOffset(), other.GetEndOffset());
 		assert(begin_off < end_off);
 
-		return FileChunkDesc(begin_off, end_off - begin_off);
+		return FileChunkDesc(begin_off, (size_t)(end_off - begin_off));
 	}
 
 	bool Contains(const FileChunkDesc& other) const
